requirements.txt
requests
pandas
py4j
pyspark
python-dotenv
google-cloud-bigquery
google-cloud-logging
google-cloud-pubsub
google-cloud-storage